1_Wikipedia_and_Math/requirements.txt
python-dotenv
langchain
langchain_core
langchain_community
langchain_google_genai
langgraph
nltk